+++ Kernel/Src/KRE/CppRTS.cpp
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-#include <KRE/CppRuntimeSupport.h>
+#include <KRE/CppRTS.h>
 
 void (*ON_CXA_PURE_VIRTUAL)() = [] {}; // NOLINT
